pyproject.toml
[project]
name = "alt-alpha-ranker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "playwright>=1.55.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
